requirements.txt
load_dotenv==0.1.0
flask==2.2.3
langchain
openai